Rename file spec.ts to VegaSpec.ts
The renaming should avoid collision with Jest tests. In the tsconfig_dev.ts it is configured to look for `spec.ts` files.
