+++ src/internal/VegaSpec.ts
